Prevent error with missing site urls in Pageflow 15.1
Since the extraction of the paged entry type into a separate engine,
calling `link_to` with nil for a URL causes an error. Rails tries to
turn an empty hash into a route. This no longer works inside the paged
engine, since it only contains editor related routes.

As a workaround, default to an empty string. This leads to the same
behavior as before: Creating a link to the current location.

Extend CI setup to make sure the site renders in cases like this.

REDMINE-17412